Make commDiags counters larger to prevent overlow in long-running programs.
In a recent long application run with commDiagnostics enabled, I saw
that one of the counters (test_nb_gets) had overflowed.  Convert the
counters from 32 bit signed integers to 64 bit unsigned ones to prevent
this from happening again in the future.


git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@20071 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ runtime/src/comm/gasnet/comm-gasnet.c
@@ -890,35 +890,35 @@ void chpl_getCommDiagnosticsHere(chpl_commDiagnostics *cd) {
   chpl_sync_unlock(&chpl_comm_diagnostics_sync);
 }
 
-int32_t chpl_numCommGets(void) {
+uint64_t chpl_numCommGets(void) {
   return chpl_comm_commDiagnostics.get;
 }
 
-int32_t chpl_numCommNBGets(void) {
+uint64_t chpl_numCommNBGets(void) {
   return chpl_comm_commDiagnostics.get_nb;
 }
 
-int32_t chpl_numCommTestNBGets(void) {
+uint64_t chpl_numCommTestNBGets(void) {
   return chpl_comm_commDiagnostics.get_nb_test;
 }
 
-int32_t chpl_numCommWaitNBGets(void) {
+uint64_t chpl_numCommWaitNBGets(void) {
   return chpl_comm_commDiagnostics.get_nb_wait;
 }
 
-int32_t chpl_numCommPuts(void) {
+uint64_t chpl_numCommPuts(void) {
   return chpl_comm_commDiagnostics.put;
 }
 
-int32_t chpl_numCommFastForks(void) {
+uint64_t chpl_numCommFastForks(void) {
   return chpl_comm_commDiagnostics.fork_fast;
 }
 
-int32_t chpl_numCommForks(void) {
+uint64_t chpl_numCommForks(void) {
   return chpl_comm_commDiagnostics.fork;
 }
 
-int32_t chpl_numCommNBForks(void) {
+uint64_t chpl_numCommNBForks(void) {
   return chpl_comm_commDiagnostics.fork_nb;
 }
